try reverting assetPrefix changes
